requirements.txt
fastapi
uvicorn
python-multipart
requests
pillow
imagehash
pytest
requests-mock